refactor: modularize map app
