+++ constants.ts
@@ -1,0 +1,118 @@
+export const USER_PROVIDED_GOOGLE_MAPS_API_KEY =
+  'AIzaSyAJPTwj4S8isr4b-3NtqVSxk450IAS1lOQ';
+
+export const DEFAULT_ITINERARY_JSON = {
+  stops: [
+    'Los Angeles, CA',
+    'Phoenix, AZ',
+    'El Paso, TX',
+    'Dallas, TX',
+    'New Orleans, LA',
+    'Atlanta, GA',
+    'Richmond, VA',
+    'New York, NY',
+    'Buffalo, NY',
+    'Chicago, IL',
+    'Minneapolis, MN',
+    'Billings, MT',
+    'Boise, ID',
+    'Reno, NV',
+    'Los Angeles, CA',
+  ],
+  dates: [
+    '2025-08-28',
+    '2025-08-29',
+    '2025-08-30',
+    '2025-08-31',
+    '2025-09-01',
+    '2025-09-02',
+    '2025-09-03',
+    '2025-09-04',
+    '2025-09-05',
+    '2025-09-06',
+    '2025-09-07',
+    '2025-09-08',
+    '2025-09-09',
+    '2025-09-10',
+    '2025-09-11',
+  ],
+  accommodations: [
+    {
+      price_per_night: 150,
+      hotel_name: 'Hilton Garden Inn Los Angeles/Hollywood',
+      address: '2005 N Highland Ave, Los Angeles, CA 90068',
+    },
+    {
+      price_per_night: 120,
+      hotel_name: 'Hyatt Place Phoenix - Downtown',
+      address: '150 W Adams St, Phoenix, AZ 85003',
+    },
+    {
+      price_per_night: 180,
+      hotel_name: 'Hotel Indigo El Paso Downtown',
+      address: '325 N Kansas St, El Paso, TX 79901',
+    },
+    {
+      price_per_night: 160,
+      hotel_name:
+        'Courtyard by Marriott Dallas Downtown/Reunion District',
+      address: '310 S Houston St, Dallas, TX 75202',
+    },
+    {
+      price_per_night: 170,
+      hotel_name: 'The Old No. 77 Hotel & Chandlery',
+      address: '535 Tchoupitoulas St, New Orleans, LA 70130',
+    },
+    {
+      price_per_night: 140,
+      hotel_name: 'Hyatt Centric Midtown Atlanta',
+      address: '125 10th St NE, Atlanta, GA 30309',
+    },
+    {
+      price_per_night: 155,
+      hotel_name: 'Graduate Richmond',
+      address: '301 W Franklin St, Richmond, VA 23220',
+    },
+    {
+      price_per_night: 180,
+      hotel_name: 'Moxy NYC Times Square',
+      address: '485 7th Ave, New York, NY 10018',
+    },
+    {
+      price_per_night: 160,
+      hotel_name: 'Aloft Buffalo Downtown',
+      address: '500 Pearl St, Buffalo, NY 14202',
+    },
+    {
+      price_per_night: 140,
+      hotel_name: 'Freehand Chicago',
+      address: '19 E Ohio St, Chicago, IL 60611',
+    },
+    {
+      price_per_night: 150,
+      hotel_name: 'Radisson RED Minneapolis Downtown',
+      address: '609 S 3rd St, Minneapolis, MN 55415',
+    },
+    {
+      price_per_night: 120,
+      hotel_name: 'Northern Hotel',
+      address: '19 N Broadway, Billings, MT 59101',
+    },
+    {
+      price_per_night: 165,
+      hotel_name: 'The Grove Hotel',
+      address: '245 S Capitol Blvd, Boise, ID 83702',
+    },
+    {
+      price_per_night: 165,
+      hotel_name: 'Renaissance Reno Downtown Hotel & Spa',
+      address: 'One S Lake St, Reno, NV 89501',
+    },
+  ],
+};
+
+export const EXAMPLE_PROMPTS: string[] = [];
+
+export const DEFAULT_GAS_PRICE_PER_GALLON = 3.5;
+export const METERS_TO_MILES = 0.000621371;
+
