+++ icons.ts
@@ -1,0 +1,69 @@
+import {html} from 'lit';
+
+export const ICON_BUSY = html`<svg
+  class="rotating"
+  xmlns="http://www.w3.org/2000/svg"
+  height="24px"
+  viewBox="0 -960 960 960"
+  width="24px"
+  fill="currentColor">
+  <path
+    d="M480-80q-82 0-155-31.5t-127.5-86Q143-252 111.5-325T80-480q0-83 31.5-155.5t86-127Q252-817 325-848.5T480-880q17 0 28.5 11.5T520-840q0 17-11.5 28.5T480-800q-133 0-226.5 93.5T160-480q0 133 93.5 226.5T480-160q133 0 226.5-93.5T800-480q0-17 11.5-28.5T840-520q17 0 28.5 11.5T880-480q0 82-31.5 155t-86 127.5q-54.5 54.5-127 86T480-80Z" />
+</svg>`;
+
+export const ICON_CHEVRON_LEFT = html`<svg
+  xmlns="http://www.w3.org/2000/svg"
+  height="24px"
+  viewBox="0 -960 960 960"
+  width="24px"
+  fill="currentColor">
+  <path
+    d="M561-240 321-480l240-240 57 57-183 183 183 183-57 57Zm-240 0L81-480l240-240 57 57-183 183 183 183-57 57Z" />
+</svg>`;
+
+export const ICON_CHEVRON_RIGHT = html`<svg
+  xmlns="http://www.w3.org/2000/svg"
+  height="24px"
+  viewBox="0 -960 960 960"
+  width="24px"
+  fill="currentColor">
+  <path
+    d="m321-80-57-57 183-183-183-183 57-57 240 240-240 240Zm240 0-57-57 183-183-183-183 57-57 240 240-240 240Z" />
+</svg>`;
+
+export const ICON_CHEVRON_UP = html`<svg
+  xmlns="http://www.w3.org/2000/svg"
+  height="24px"
+  viewBox="0 -960 960 960"
+  width="24px"
+  fill="currentColor">
+  <path d="M480-584 240-344l-56-56 296-296 296 296-56 56-240-240Z" />
+</svg>`;
+
+export const ICON_CHEVRON_DOWN = html`<svg
+  xmlns="http://www.w3.org/2000/svg"
+  height="24px"
+  viewBox="0 -960 960 960"
+  width="24px"
+  fill="currentColor">
+  <path d="M480-344 240-584l56-56 184 184 184-184 56 56-240 240Z" />
+</svg>`;
+
+export const ICON_PLAY = html`<svg
+  xmlns="http://www.w3.org/2000/svg"
+  height="24px"
+  viewBox="0 -960 960 960"
+  width="24px"
+  fill="currentColor">
+  <path d="M320-200v-560l440 280-440 280Z" />
+</svg>`;
+
+export const ICON_STOP = html`<svg
+  xmlns="http://www.w3.org/2000/svg"
+  height="24px"
+  viewBox="0 -960 960 960"
+  width="24px"
+  fill="currentColor">
+  <path d="M280-280h400v-400H280v400Z" />
+</svg>`;
+
